Split out AffectedProjectsComputer for testability
